refactor!: Filter to Contracts namespace
